rename to replace the word page with screen to keep consistent
