+++ lib/presentation/providers/providers.dart
@@ -6,10 +6,10 @@ import 'package:flap_app/data/source/network/spoonacular_api_impl.dart';
 import 'package:flap_app/domain/repository/auth/auth_repository.dart';
 import 'package:flap_app/domain/repository/recipe/recipe_repository.dart';
 import 'package:flap_app/domain/usecase/log_in_usecase.dart';
-import 'package:flap_app/presentation/screens/home/notifier/home_page_state.dart';
-import 'package:flap_app/presentation/screens/home/notifier/home_page_state_notifier.dart';
-import 'package:flap_app/presentation/screens/login/notifier/login_page_state.dart';
-import 'package:flap_app/presentation/screens/login/notifier/login_page_state_notifier.dart';
+import 'package:flap_app/presentation/screens/home/notifier/home_screen_state.dart';
+import 'package:flap_app/presentation/screens/home/notifier/home_screen_state_notifier.dart';
+import 'package:flap_app/presentation/screens/login/notifier/login_screen_state.dart';
+import 'package:flap_app/presentation/screens/login/notifier/login_screen_state_notifier.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 //TODO: Investigate dependency injection solutions for passing in Dio to api class
