+++ lib/presentation/screens/home/notifier/home_screen_state_notifier.dart
@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:flap_app/domain/entity/recipe.dart';
 import 'package:flap_app/domain/entity/request_response.dart';
 import 'package:flap_app/domain/repository/recipe/recipe_repository.dart';
-import 'package:flap_app/presentation/screens/home/notifier/home_page_state.dart';
+import 'package:flap_app/presentation/screens/home/notifier/home_screen_state.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class HomePageStateNotifier extends StateNotifier<HomePageState> {
