+++ lib/presentation/screens/login/notifier/login_screen_state.dart
@@ -60,12 +60,4 @@ class LoginScreenState {
         );
     }
   }
-
-  // bool isLoading() {
-  //   return loginEvent is LoadingEvent;
-  // }
-
-  // bool isSuccess() {
-  //   return loginEvent is EventSuccess;
-  // }
 }
