+++ lib/presentation/screens/login/notifier/login_screen_state_notifier.dart
@@ -4,7 +4,7 @@ import 'package:flap_app/domain/entity/login_info.dart';
 import 'package:flap_app/domain/entity/request_response.dart';
 import 'package:flap_app/domain/entity/user_info.dart';
 import 'package:flap_app/domain/usecase/log_in_usecase.dart';
-import 'package:flap_app/presentation/screens/login/notifier/login_page_state.dart';
+import 'package:flap_app/presentation/screens/login/notifier/login_screen_state.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class LoginPageStateNotifier extends StateNotifier<LoginScreenState> {
